Refactor 01_schema and seeds and Add db queries

+++ db/schema/01_schema.sql
@@ -17,7 +17,7 @@ CREATE TABLE menu_items (
   id SERIAL PRIMARY KEY NOT NULL,
   name VARCHAR(255) NOT NULL,
   description text,
-  price INTEGER NOT NULL DEFAULT 0,
+  price money NOT NULL DEFAULT 0,
   availability BOOLEAN NOT NULL DEFAULT TRUE,
   image_url VARCHAR(255)
 );
@@ -38,7 +38,7 @@ CREATE TABLE orders (
   id SERIAL PRIMARY KEY NOT NULL,
   customer_id INTEGER REFERENCES users(id) ON DELETE CASCADE,
   status INTEGER REFERENCES status(id) ON DELETE CASCADE,
-  total INTEGER NOT NULL DEFAULT 0,
+  total money NOT NULL DEFAULT 0,
   created_at TIMESTAMP DEFAULT NOW(),
   accepted_at DATE,
   estimated_end_time DATE,
